requirements.txt
requests
beautifulsoup4
pandas
fake_useragent
duckduckgo_search
python-dotenv
openai
streamlit
plotly